+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [v1.0.1] - 2018-11-03
 
-Patch: Update README with more specific information.
+Patch: 
+- Update README with more specific information.
 
 ## [v1.0.0] - 2018-11-02