+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## unreleased
+
+Feature:
+- Rename the module `is-plain-number` to `is-string-a-number`: after some
+  thoughts, I realised that the purpose of my module was to specifically check
+  if a `string` value was representing a number in decimal base and notation.
+
 ## [v1.0.1] - 2018-11-03
 
-Patch: 
+Patch:
 - Update README with more specific information.
 
 ## [v1.0.0] - 2018-11-02